refactor: reorganize imports and initialize service registry
